[VPR] Remove the dead loop between rr_graph2.h and rr_graph_storage.h; Move the common data structure to rr_edge.h; rr_graph2.h is no longer a dependency on rr_graph_storage.h which is supposed to be the lowest level of head files
