+++ vpr/src/route/rr_graph.cpp
@@ -573,7 +573,7 @@ static void build_rr_graph(const t_graph_type graph_type,
     /* Alloc node lookups, count nodes, alloc rr nodes */
     int num_rr_nodes = 0;
 
-    alloc_and_load_rr_node_indices(device_ctx.rr_node_indices,
+    alloc_and_load_rr_node_indices(device_ctx.rr_graph_builder,
                                    max_chan_width, grid,
                                    &num_rr_nodes, chan_details_x, chan_details_y);
