+++ vpr/src/route/rr_graph_storage.h
@@ -6,12 +6,13 @@
 
 #include "rr_graph_fwd.h"
 #include "rr_node_fwd.h"
-#include "rr_graph2.h"
+#include "rr_edge.h"
 #include "vtr_log.h"
 #include "vtr_memory.h"
 #include "vpr_utils.h"
 #include "vtr_strong_id_range.h"
 #include "vtr_array_view.h"
+
 
 /* Main structure describing one routing resource node.  Everything in       *
  * this structure should describe the graph -- information needed only       *
